refactor: extract mac traits
